organized the actions
