+++ src/main/java/frc/robot/actions/elevatorActions/MoveToUpperCone.java
@@ -1,15 +1,15 @@
-package frc.robot.actions;
+package frc.robot.actions.elevatorActions;
 
 import com.flash3388.flashlib.scheduling.ActionControl;
 import com.flash3388.flashlib.scheduling.FinishReason;
 import com.flash3388.flashlib.scheduling.actions.ActionBase;
 import frc.robot.subSystems.ElevatorSystem;
 
-public class MoveToMiddleCone extends ActionBase {
+public class MoveToUpperCone extends ActionBase {
     private ElevatorSystem elevatorSystem;
 
 
-    public MoveToMiddleCone(ElevatorSystem elevatorSystem){
+    public MoveToUpperCone(ElevatorSystem elevatorSystem){
         this.elevatorSystem = elevatorSystem;
         requires(this.elevatorSystem);
     }
@@ -21,12 +21,12 @@ public class MoveToMiddleCone extends ActionBase {
 
     @Override
     public void execute(ActionControl control) {
-        elevatorSystem.moveToMiddleCone();
+        elevatorSystem.moveToUpperCone();
     }
 
     @Override
     public boolean isFinished(){
-        return elevatorSystem.isAtMiddleCone();
+        return elevatorSystem.isAtUpperCone();
     }
 
     @Override
